0.0.1 - first public release
WebLamp is my first attempt to use a ESP8266 module.
With this code you can build a web enabled lamp.
The lamp connects at startup to a wifi network (with static or dinamic
ip) and exposes a mini webs erver on port 80 to drive simple REST
commands
See the code for details